NTT: parameterisation doesn't work. Copy and optimise implementations
